Fixed #34112 -- Added async interface to model methods
