+++ tests/async/models.py
@@ -1,5 +1,16 @@
 from django.db import models
+from django.utils import timezone
+
+
+class RelatedModel(models.Model):
+    simple = models.ForeignKey("SimpleModel", models.CASCADE, null=True)
 
 
 class SimpleModel(models.Model):
     field = models.IntegerField()
+    created = models.DateTimeField(default=timezone.now)
+
+
+class Book(models.Model):
+    title = models.CharField(max_length=200)
+    pages = models.IntegerField(default=0)
