+++ tests/async/test_async_queryset.py
@@ -243,6 +243,6 @@ class AsyncQuerySetTest(TestCase):
                         self.fail(f"QuerySet.aexplain() result is not valid JSON: {e}")
 
     async def test_raw(self):
-        sql = "SELECT id, field FROM async_queryset_simplemodel WHERE created=%s"
+        sql = "SELECT id, field FROM async_simplemodel WHERE created=%s"
         qs = SimpleModel.objects.raw(sql, [self.s1.created])
         self.assertEqual([o async for o in qs], [self.s1])
