--- tests/async_queryset/models.py
@@ -1,11 +1,0 @@
-from django.db import models
-from django.utils import timezone
-
-
-class RelatedModel(models.Model):
-    simple = models.ForeignKey("SimpleModel", models.CASCADE, null=True)
-
-
-class SimpleModel(models.Model):
-    field = models.IntegerField()
-    created = models.DateTimeField(default=timezone.now)
